requirements.txt
py-clob-client>=0.34.0
websockets>=12.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.26.0
ccxt>=4.2.0
ta>=0.11.0
python-dotenv>=1.0.0
pyyaml>=6.0.0

